+++ src/wenmode/plugins/cjk_friendly.py
@@ -9,7 +9,7 @@ if TYPE_CHECKING:
     from wenmode import Wenmode
 
 
-def setup(wenmode: Wenmode, **options: Any) -> None:
-    wenmode.register_rule(Emphasis(cjk_friendly=True))
-    if 'extended_autolink' in wenmode.parser.rules:
-        wenmode.register_rule(ExtendedAutolink(cjk_friendly=True))
+def setup(wen: Wenmode, **options: Any) -> None:
+    wen.register_rule(Emphasis(cjk_friendly=True))
+    if 'extended_autolink' in wen.parser.rules:
+        wen.register_rule(ExtendedAutolink(cjk_friendly=True))